requirements.txt
tensorflow
numpy
opencv-python-headless
Pillow
streamlit